+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added diagnostic area _Quality_, _Support_ and _Requirement_ 
 * Added Packages module to report installed packages and dependencies
 * Fixed analysis platform on incremental audit
+* Fixed compilation error due to newer com.unity.nuget.mono-cecil 
 * Improved text search to match custom properties
 
 ## [0.8.3-preview] - 2022-09-05