+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 * Added 'documentation' support to descriptor
-* Added diagnostic area _Quality_, _Support_ and _Requirement_ 
+* Added diagnostic area _Quality_, _Support_ and _Requirement_
+* Added Compute Shader Variants support
 * Added Packages module to report installed packages and dependencies
 * Fixed analysis platform on incremental audit
 * Fixed compilation error due to newer com.unity.nuget.mono-cecil