+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added Packages module to report installed packages and dependencies
 * Fixed analysis platform on incremental audit
 * Fixed compilation error due to newer com.unity.nuget.mono-cecil 
+* Fixed package version parsing exception
 * Improved text search to match custom properties
 
 ## [0.8.3-preview] - 2022-09-05