+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.4-preview] - 2022-09-27
+* Fixed compilation error due to newer com.unity.nuget.mono-cecil 
+
 ## [0.8.3-preview] - 2022-09-05
 * Fixed *NullReferenceException* on Draw2D shader not being found
 * Added HTML export support